refactor(Gallery): move styles
